+++ packages/extension-browser/tests/unit/background-gmail-silent-failure.test.ts
@@ -98,6 +98,7 @@ function baseProfile(): ExtensionProfile {
     cloudProvider: null,
     cloudModel: null,
     cloudApiKey: null,
+    cloudApiKeys: {},
     historyEnabled: true,
     displayName: "you",
     neurotypes: [],