+++ packages/extension-browser/tests/unit/background-handlers.test.ts
@@ -66,6 +66,7 @@ function baseProfile(
     cloudProvider: null,
     cloudModel: null,
     cloudApiKey: null,
+    cloudApiKeys: {},
     historyEnabled: true,
     displayName: "you",
     neurotypes: [],